+++ packages/tailwind/src/tailwind-presets.ts
@@ -1,8 +1,6 @@
 import { ThemeConfig } from "tailwindcss/types/config"
 import { defaultTheme } from "@kaizen/design-tokens"
-import { kzHeight } from "./kz-height"
 import { kzSpacing } from "./kz-spacing"
-import { kzWidth } from "./kz-width"
 
 export type KaizenTailwindTheme = Partial<ThemeConfig>
 export interface KaizenTailwindPreset {
@@ -97,28 +95,10 @@ export const kaizenTailwindTheme: KaizenTailwindTheme = {
     "letter-spacing-normal": "normal",
   },
   // A mix of layout styles
-  maxWidth: kzWidth,
-  width: kzWidth,
-  maxHeight: kzHeight,
-  height: kzHeight,
   screens: {
     md: defaultTheme.layout.breakpoints.medium, // => @media (min-width: 768px) { ... }
     lg: defaultTheme.layout.breakpoints.large, // => @media (min-width: 1080px) { ... }
-    "md-max": { max: defaultTheme.layout.breakpoints.medium }, // => @media (max-width: 768px) { ... }
-    "lg-max": { max: defaultTheme.layout.breakpoints.large }, // => @media (min-width: 1080px) { ... }
   },
-  // TODO: These work in TW Play but not in our Storybook config.
-  //        Add these when we get them working in Storybook.
-  // ringWidth: {
-  //   "default-width": `${defaultTheme.border.focusRing.borderWidth}`,
-  // },
-  // ringOffsetWidth: {
-  //   "default-width": "2px",
-  // },
-  // ringColor: {
-  //   "default-color": `${defaultTheme.color.blue[500]}`,
-  // },
-  // ringOffsetColor: {},
 }
 
 export const Preset: KaizenTailwindPreset = {