+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 10.3.7
+
+### Patch Changes
+
+- [#3608](https://github.com/cultureamp/kaizen-design-system/pull/3608) [`cc96d9835`](https://github.com/cultureamp/kaizen-design-system/commit/cc96d98351ae380fe1a1e8a33d65d80232ed7a57) - Update usages of classnames to not use object syntax
+
 ## 10.3.6
 
 ### Patch Changes